requirements.txt
django
djangorestframework
pillow
django-filter